% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSMHelper.R
\name{candidate.detfcn.summary}
\alias{candidate.detfcn.summary}
\title{Produce a summary table of candidate detection functions for one DDF spec}
\usage{
candidate.detfcn.summary(df.spec.name, species)
}
\arguments{
\item{df.spec.name}{Character string DDF spec name (used to locate the
results folder and label the output).}

\item{species}{Character string species code.}
}
\value{
Data frame of candidate model statistics sorted by AIC.
}
\description{
Produce a summary table of candidate detection functions for one DDF spec
}
