% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.dataset}
\alias{get.dataset}
\title{Extract the dataset component from a DDF spec name}
\usage{
get.dataset(nm)
}
\arguments{
\item{nm}{Character string DDF spec name (e.g. \code{"ECSAS_Ship_F_D"}).}
}
\value{
Character string dataset identifier (e.g. \code{"ECSAS"}).
}
\description{
Extract the dataset component from a DDF spec name
}
