% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSMHelper.R
\name{get.platform.class}
\alias{get.platform.class}
\title{Extract the platform class from a DDF spec name}
\usage{
get.platform.class(nm)
}
\arguments{
\item{nm}{Character string DDF spec name.}
}
\value{
\code{"Ship"} or \code{"Aerial"}.
}
\description{
Extract the platform class from a DDF spec name
}
