% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSMHelper.R
\name{make.line}
\alias{make.line}
\title{Make an sf linestring from a two-row coordinate matrix}
\usage{
make.line(xy2)
}
\arguments{
\item{xy2}{Numeric matrix with 2 rows and 2 columns (lon, lat).}
}
\value{
An \code{sfg} linestring object.
}
\description{
Make an sf linestring from a two-row coordinate matrix
}
