% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.dens.est}
\alias{print.dens.est}
\title{Pretty-print the output of get.dens.est}
\usage{
\method{print}{dens.est}(densEst)
}
\arguments{
\item{densEst}{Named list as returned by \code{\link{get.dens.est}}.}
}
\value{
\code{invisible(NULL)}, called for its side-effect (printed output).
}
\description{
\strong{Note: not currently used.}
}
