% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{registered.dsm.families}
\alias{registered.dsm.families}
\title{Keys of the response families currently registered}
\usage{
registered.dsm.families()
}
\value{
Character vector of keys, sorted.
}
\description{
Keys of the response families currently registered
}
\examples{
registered.dsm.families()
}
\seealso{
\code{\link{register.dsm.family}}
}
