% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove.dsm.summaries}
\alias{remove.dsm.summaries}
\title{Delete all DSM summary RData files for all species groups}
\usage{
remove.dsm.summaries()
}
\value{
\code{invisible(NULL)}, called for its side-effect (files deleted).
}
\description{
Delete all DSM summary RData files for all species groups
}
