% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{watches.to.lines}
\alias{watches.to.lines}
\title{Convert a watches data frame to a SpatialLinesDataFrame}
\usage{
watches.to.lines(watches)
}
\arguments{
\item{watches}{Data frame with columns \code{WatchID}, \code{LongStart},
\code{LatStart}, \code{LongEnd}, \code{LatEnd}.}
}
\value{
\code{SpatialLinesDataFrame} in WGS 84 (EPSG 4326).
}
\description{
Convert a watches data frame to a SpatialLinesDataFrame
}
