% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSMHelper.R
\name{create.dir.if.needed}
\alias{create.dir.if.needed}
\title{Create a directory recursively if it does not already exist}
\usage{
create.dir.if.needed(pathname)
}
\arguments{
\item{pathname}{Character string directory path to create.}
}
\value{
\code{pathname} (invisibly).
}
\description{
Returns \code{pathname} invisibly so the function can be used in a
pipeline.
}
