% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSMHelper.R
\name{get.behav}
\alias{get.behav}
\title{Extract the behaviour component from a DDF spec name}
\usage{
get.behav(nm)
}
\arguments{
\item{nm}{Character string DDF spec name.}
}
\value{
\code{"F"} (flying) or \code{"W"} (water).
}
\description{
Extract the behaviour component from a DDF spec name
}
