% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.dist_type}
\alias{get.dist_type}
\title{Extract the distance-type component from a DDF spec name}
\usage{
get.dist_type(nm)
}
\arguments{
\item{nm}{Character string DDF spec name.}
}
\value{
\code{"D"} (distances available) or \code{"N"} (no distances).
}
\description{
Extract the distance-type component from a DDF spec name
}
