% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set.def.df.spec.values}
\alias{set.def.df.spec.values}
\title{Set default detection function spec values by platform class}
\usage{
set.def.df.spec.values(ddf.def, nm)
}
\arguments{
\item{ddf.def}{A single DDF spec list to be populated.}

\item{nm}{Character string DDF spec name of the form
\code{"<dataset>_<Ship|Aerial>_<behav>_<D|N>"}.}
}
\value{
\code{ddf.def} with platform-appropriate default values filled in.
}
\description{
Populates \code{convert_units}, \code{cutpoints}, and
\code{distance.centers} in a DDF spec list based on whether the spec name
indicates a ship or aerial survey.
}
