% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize.dsm}
\alias{summarize.dsm}
\title{Produce a one-row summary data frame for a DSM model}
\usage{
summarize.dsm(model)
}
\arguments{
\item{model}{A fitted \code{dsm} object, or a \code{try-error} if model
fitting failed.}
}
\value{
Single-row data frame with columns \code{response}, \code{terms},
\code{AIC}, \code{REML}, \code{OverDisp}, and \code{Deviance_explained}.
}
\description{
Produce a one-row summary data frame for a DSM model
}
