% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{watch.map}
\alias{watch.map}
\title{Quick leaflet map of watch start positions for debugging}
\usage{
watch.map(dat)
}
\arguments{
\item{dat}{Data frame with \code{LongStart}, \code{LatStart}, and
\code{WatchID} columns.}
}
\value{
A \code{leaflet} map object.
}
\description{
Quick leaflet map of watch start positions for debugging
}
